requirements.txt
requests>=2.25.0
pandas>=1.3.0
numpy>=1.20.0
websocket-client>=1.0.0
pyyaml>=5.4.0
colorama>=0.4.4
